+++ lib/flutter_musical_notation.dart
@@ -1,3 +1,4 @@
+export 'src/measure_accidentals.dart';
 export 'src/music_notation.dart';
 export 'src/music_notation_painter.dart';
 export 'src/notation_layout.dart';